+++ server/models/Order.ts
@@ -1,4 +1,5 @@
 import z from 'zod'
+import AddressModel from './Address'
 import AlcoholMonitoringModel from './AlcoholMonitoring'
 import AttachmentModel from './Attachment'
 import AttendanceMonitoringModel from './AttendanceMonitoring'
@@ -7,7 +8,6 @@ import CurfewConditionsModel from './CurfewConditions'
 import CurfewReleaseDateModel from './CurfewReleaseDate'
 import CurfewTimetableModel from './CurfewTimetable'
 import DeviceWearerModel from './DeviceWearer'
-import DeviceWearerAddressModel from './DeviceWearerAddress'
 import DeviceWearerResponsibleAdultModel from './DeviceWearerResponsibleAdult'
 import EnforcementZoneModel from './EnforcementZone'
 import InstallationAndRiskModel from './InstallationAndRisk'
@@ -20,7 +20,7 @@ const OrderModel = z.object({
   id: z.string().uuid(),
   status: OrderStatusEnum,
   deviceWearer: DeviceWearerModel,
-  deviceWearerAddresses: z.array(DeviceWearerAddressModel),
+  addresses: z.array(AddressModel),
   deviceWearerResponsibleAdult: DeviceWearerResponsibleAdultModel.nullable(),
   deviceWearerContactDetails: DeviceWearerContactDetailsModel,
   enforcementZoneConditions: z.array(EnforcementZoneModel),
