+++ server/services/index.ts
@@ -15,7 +15,7 @@ import MonitoringConditionsService from './monitoringConditionsService'
 import OrderSearchService from './orderSearchService'
 import OrderService from './orderService'
 import TrailMonitoringService from './trailMonitoringService'
-import DeviceWearerAddressService from './deviceWearerAddressService'
+import AddressService from './addressService'
 
 export const services = () => {
   const { applicationInfo, hmppsAuditClient, cemoApiClient } = dataAccess()
@@ -28,7 +28,7 @@ export const services = () => {
   const curfewConditionsService = new CurfewConditionsService(cemoApiClient)
   const curfewReleaseDateService = new CurfewReleaseDateService(cemoApiClient)
   const curfewTimetableService = new CurfewTimetableService(cemoApiClient)
-  const deviceWearerAddressService = new DeviceWearerAddressService(cemoApiClient)
+  const addressService = new AddressService(cemoApiClient)
   const deviceWearerResponsibleAdultService = new DeviceWearerResponsibleAdultService(cemoApiClient)
   const deviceWearerService = new DeviceWearerService(cemoApiClient)
   const installationAndRiskService = new InstallationAndRiskService(cemoApiClient)
@@ -48,7 +48,7 @@ export const services = () => {
     curfewReleaseDateService,
     curfewConditionsService,
     curfewTimetableService,
-    deviceWearerAddressService,
+    addressService,
     deviceWearerResponsibleAdultService,
     deviceWearerService,
     installationAndRiskService,
